Update action to use official AWS Actions (#1)
### Why:
This action uses a docker container managed by a Utrust former
colleague. While we do still trust him, the image has not been updated
in a few years, and having an important piece of our CI flow out of our
hands is not intended.

### This addresses the issue by:
Removing naps62 docker image dependency and replace it with AWS official
action to sign in to ECR and keep the script to push new docker image to
the right place.